canary-backend/requirements.txt
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
urllib3==2.5.0
PyJWT==2.8.0
# Note: The versions specified are examples and may need to be updated based on your requirements.
# Ensure that the versions are compatible with your project and AWS SDK requirements.